feature (userContext): with firebase
