+++ packages/client/src/containers/HomePage/HomePage.Container.js
@@ -1,7 +1,6 @@
 import React from 'react';
 
 import './HomePage.styles.css';
-
 import { Header } from '../../components/Header/Header.component';
 import { Body } from '../../components/Body/Body.component';
 import Newsletter from '../../components/Newsletter/Newsletter.component';
