+++ packages/client/src/containers/Login/Login.container.js
@@ -1,17 +1,13 @@
 import React, { useEffect, useState } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
-import {
-  auth,
-  logInWithEmailAndPassword,
-  signInWithGoogle,
-} from '../../firebase';
-import { useAuthState } from 'react-firebase-hooks/auth';
+import { useUser } from '../../userContext';
 import './Login.styles.css';
 
 function Login() {
+  const { logInWithEmailAndPassword, signInWithGoogle, loading, user } =
+    useUser();
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
-  const [user, loading] = useAuthState(auth);
   const navigate = useNavigate();
   useEffect(() => {
     if (loading) {
