+++ packages/client/src/containers/Signup/Signup.container.js
@@ -1,18 +1,14 @@
 import React, { useEffect, useState } from 'react';
-import { useAuthState } from 'react-firebase-hooks/auth';
 import { Link, useNavigate } from 'react-router-dom';
-import {
-  auth,
-  registerWithEmailAndPassword,
-  signInWithGoogle,
-} from '../../firebase';
+import { useUser } from '../../userContext';
 import './Signup.styles.css';
 
 function Signup() {
+  const { user, loading, registerWithEmailAndPassword, signInWithGoogle } =
+    useUser();
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
   const [name, setName] = useState('');
-  const [user, loading] = useAuthState(auth);
   const navigate = useNavigate();
   const register = () => {
     /* if (!name) alert('Please enter name'); add some error message & validation */
